Simplify toolwatcher
At the moment, toolwatcher uses a cascade of grep, cut and two levels
of sed to extract a list of all tools.  This change simplifies that to
a single sed regular expression.

Also, permissions for the newly created directories are set by mkdir's
"-m" option instead of by a subsequent chmod call.  This removes a
(very theoretically exploitable) race condition as well.

Change-Id: I7a79c8fb6d1727e8d49febd964a4e1075e48a688