Remove rmtool
rmtool has not been used for some time and shows this by inter alia
only archiving user databases on tools-db, but not the replica
servers, referring to the webservers file, etc.  Additionally, I
personally do not like using it automatically because for simple
deletions of tools it is overly complicated, while in more complex
cases it might not do The Right Thing™.  So this change removes the
script and its man page.

Change-Id: I88159ebc6961cb169b74ab5d2c6e5f9a36fe7df0