Let take fail if recursion failed
Currently, calling take on a directory succeeds if the directory
itself could be assumed ownership of regardless of its children.
While the logic to sum up the results of the recursion was in the
source code, its result was ignored.

This change implements the common behaviour for a program to only
succeed if there were no errors at all.

Change-Id: I3415bf8ee0643a1f16191426b804684680878a20