www: Title encode usernames by converting space to '_'
Php's urlencode() converts spaces to '+'. Wikitech 301 redirects an URL
like https://wikitech.wikimedia.org/wiki/User:Foo+Bar to
https://wikitech.wikimedia.org/wiki/User:Foo%2BBar. This happens because
urlencode() generates application/x-www-form-urlencoded strings which
are technically only valid in a query string or form data. RFC 3986
compliant URIs must use '%20' to encode a space.

There are a couple of was to fix this. One is to use rawurlencode()
which is RFC 3986 compliant. The other is to convert spaces to '_'
(MediaWiki Title's canonicalization of spaces). Using the MW encoding
should save the user a 301 redirect so lets us that method.

Change-Id: I4d944fd49b69a3bf79e2b8448de7e630d6011de8