jsub: Fix #!... to actually work
Follow up to a53fc1a2dfbf that should actually work. There is no posix
standard for processing #!... lines and my prior attempt fails miserably
on Ubuntu/Linux. The runtime behavior on most (all?) Linux versions is
to split the line on the first whitespace and treat the two parts as the
path to an interpreter and a single argument to pass it. This means that
"/usr/bin/env python2 -E -s" is parsed as ("/usr/bin/env", "python2 -E
-s"). /usr/bin/env then looks for an interpreter named 'python2 -E -s'
and fails for fairly obvious reasons.

Since the busted deb version has never been built and deployed we don't
need a version bump for the package.

Bug: T147350
Change-Id: Ia0e88d9102f5f1df6b32396913077a72833d8f62