Do not hardcore database hosts in list-user-databases
The database host that served c2.labsdb was decommissioned and is no
longer accessible.  This causes list-user-databases to fail because
c2.labsdb is in the hardcoded list of database servers to check for a
user's databases.

There is no easy authoritative source for a list of active database
servers; role::labs::db::maintain_dbusers lists all database servers,
but does not differentiate between hosts that are accessible from the
Labs network and hosts that are firewalled.  Therefore this change
removes the hardcoded list completely and instead requires that
list-user-databases is called with an explicit list of database
servers to check for a user's database.

For the active database servers current at the time of this change,
list-user-databases would be invoked as:

| $ list-user-databases -H labsdb1001.eqiad.wmnet \
|                       -H labsdb1003.eqiad.wmnet \
|                       -H labsdb1005.eqiad.wmnet \
|                       /path/to/replica.my.cnf

Change-Id: I861375502a04a839db69fa1b92e4acfecf6470e1