crontab: convert to mostly binary processing, and use surrogates for text
Argh... Python's unicode vs binary is always the annoying part.

The the code path processing the crontab data should now be binary
only, with the exception of the part that adds jsub to the entries,
which escapes and unescapes the data with surrogates before and after
the processing. Reading and writing to stdio in binary mode is done
on the buffer of sys.std{in,out,err}

Bug: T181948
Change-Id: I7601dae42724c4abc38c4c5cca76e3ad505b5d50