jsub: recognize HOME environment variable
The job outputs are written by default under the home directory of the
user as defined in the password database. When building the Debian
package with pbuilder, the build user intentionally has a home directory
set to a non exitent directory: /nonexitent. The test suite then fails
to write to it and the build fail.

Change jsub to also recognize the HOME environment variable which, when
set, takes precedence over the entry from the password database.
That is similar to the effective_username() function just above which
recognizes the LOGNAME and USER environment variables.

In debian/rules, create a dummy homedirectory and set HOME to run the
tests.

Bug: T239910
Change-Id: I59af69faa5f5995256a1532122dfb37f5747028a