oge-crontab: Add -cwd during jsub-adding if cd is prefixed
The docblock for add_jsub() reminds developers, but not end users, that
using `cd <whatever>` in a crontab line will execute on the cron server
rather than on the job grid. This quite often will end up not doing what
the user really hoped. There is however a `-cwd` flag for jsub that will
capture the current working directory of the process that jsub is called
from and pass that on to the grid. When turning a crontab string that
does not use jsub into a grid command, we will now add `-cwd` to the
jsub argument list if a `cd <whatever>` subcommand is also seen in the
command.

Also, update tox.ini with python 3.5 in hopes that we can still target
Stretch's python.

Change-Id: Id0af7e2184a6f575d3ec5b9ca8a29054cdb7f3ac