Default to utf8mb4 to properly decode strings
By default, this wrapper was connecting with the utf8 character set,
which did not allow selecting 4-byte unicode strings (such as emojis)

NOTE: this is just a patch, you may want to change the replica.my.cnf
file that's generated for each user or take some other approach
server-side.  Or if for some reason folks want to connect with other
character sets you may want to parameterize it.

Bug: T284623
Change-Id: I0dc01fd64ab84af16a7fa883b760d165ed9e7d64